# packages/api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: Lark Radar Data Service
  version: 0.1.0
  description: |
    The local HTTP API exposed by `apps/data-service` (Go) and consumed by
    `apps/web` (Next.js). This document is the single source of truth — both
    the TypeScript types in `apps/web` and the Go types in
    `apps/data-service/api/types.gen.go` are generated from this file.

    Conventions:
      * Every response is wrapped in an envelope `{ ok: bool, ... }`. On
        `ok: false`, an `error` field is present with a human-readable
        message; other fields may also be present alongside the error.
      * Times in the wire are either ISO-ish date strings (`YYYY-MM-DD`),
        Unix epoch seconds in `timestamp` / `last_synced_at` fields, or
        opaque display strings in `time`.
      * The `Sidecar/Source/Range` query parameters are forgiving — invalid
        values fall back to defaults rather than 400.
servers:
  - url: http://localhost:3456
    description: Local Go data service (default port)
paths:
...
  /api/topics/links:
    get:
      summary: Extract links from messages on a date
      operationId: getTopicLinks
      parameters:
        - { in: query, name: date, required: true, schema: { type: string, format: date } }
      responses:
        "200":
          description: Links
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TopicLinksResponse" }
...
  /api/links:
    get:
      summary: List previously detected links (capped at 200)
      operationId: getLinks
      parameters:
        - { in: query, name: chatroom_id, schema: { type: string } }
        - { in: query, name: date, schema: { type: string, format: date } }
      responses:
        "200":
          description: Links
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LinksResponse" }
...
  /api/lark/messages:
    get:
      summary: Raw passthrough of `lark-cli im +chat-messages-list`
      description: |
        Whatever shape lark-cli emits for the given chat — the data service does
        not normalize or wrap it. Callers should treat the body as opaque JSON.
      operationId: getLarkMessages
      parameters:
        - { in: query, name: chat_id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Raw lark-cli JSON (shape varies)
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "400":
          description: Missing chat_id
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        ok: { type: boolean }
        error: { type: string }
      required:
        - ok
        - error
...
    TopicLinksResponse:
      type: object
      properties:
        ok: { type: boolean }
        date: { type: string }
        links: { type: array, items: { $ref: "#/components/schemas/Link" } }
        count: { type: integer }
      required:
        - ok
        - links
        - count
...
    Link:
      type: object
      properties:
        id: { type: integer }
        chatroom_id: { type: string }
        message_id: { type: string }
        url: { type: string }
        title: { type: string }
        date: { type: string }
        created_at: { type: string }
      required:
        - id
        - chatroom_id
        - message_id
        - url
        - date
    LinksResponse:
      type: object
      properties:
        ok: { type: boolean }
        links: { type: array, items: { $ref: "#/components/schemas/Link" } }
      required:
        - ok
        - links
...